Started changing print_exp to long double

+++ includes/ft_printf_utils.h
@@ -6,7 +6,7 @@
 /*   By: dda-silv <dda-silv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/20 15:39:35 by dda-silv          #+#    #+#             */
-/*   Updated: 2021/01/25 08:19:10 by dda-silv         ###   ########.fr       */
+/*   Updated: 2021/02/18 17:07:48 by dda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	add_sign(char **nb, char *flags);
 void	add_hex_prefix(char **nb, char type, char *flags);
 int		is_all_zeros(char *nb);
 void	add_point(char **nb, t_format *settings);
-void	get_exp_str(double *nb, char **exponent);
+void	get_exp_str(long double *nb, char **exponent);
 void	add_exponent(char **nb, char **exponent);
 void	adjust_rounding(char *nb);
 void	increment(char *nb, char sign);
